LayoutTests/platform/ios/ios/fast/forms/autocapitalize-null-expected.txt

Test null handling on the autocapitalize attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS form.getAttribute('autocapitalize') is null
form.autocapitalize = null
PASS form.getAttribute('autocapitalize') is ""
PASS form.autocapitalize is ""
PASS input.getAttribute('autocapitalize') is null
input.autocapitalize = null
PASS input.getAttribute('autocapitalize') is ""
PASS input.autocapitalize is ""
PASS textarea.getAttribute('autocapitalize') is null
textarea.autocapitalize = null
PASS textarea.getAttribute('autocapitalize') is ""
PASS textarea.autocapitalize is ""
PASS successfullyParsed is true

TEST COMPLETE
